Apply compatibility update
Updated Framework support for the core library so that it will now support standard 2.0 for the widest compatibility with .NET projects.
Removed the Dependency Injection reference so projects that reference the Mapper aren't required to use Dependency Injection.
Added extension methods to increase ease of use with the product - you can now create IMapperService directly from a MapperBuilder.
Created Benchmarks to display the performance difference between this product and Mapster and AutoMapper.
Create the Clone extension method so that the intention of cloning can be made more obvious as opposed to the usual mapping.