Context Mapping Update
Introduced context injected mapping so third-party data can be mapped alongside the mapping process. This will be useful in converting data and getting data from a database (collecting all the data required ahead of mapping of course). This has been implemented on the surface level and another update will have to be done to have the reflection extension methods support it as well as the tester.