Apply missing methods
Applied the missing MapAndInject methods that should have been present in the MapperBuilder so they can be nested as well.
Applied my favourite method, the Configure extension method, to context-aware mappings for better readability. And in that, also created a reflection based AddAssignment extension method for context-aware mappings.
Added the Configure extension method to the README because I really felt that it should mention such a great method for creating really readable mapping configurations.